Fix a bug in node ID generation - they now match IDs as given by the go tools
Was using blake2s256; they're using blake2b with output size of 32 bytes (256 bits).
I've got it using a staticly known output size at the moment. That could change in the future, but I'm fine with leaving it this way since I'm going to add a config version key and better error handling on creating the hash next.