Add a format_version key to the config; discover confy doesn't handle graceful updating from the default trait
(unless the key is literally supposed to be version, like from the example). This could be a problem much later if I need to break the format; looks like I'll have to manually fix what's wrong programatically rather than confy doing it. Unless I find a lib that does that, or get confy to.