Work on gen-node a bit: it now asks for confirmation before overwriting your config file if it exists
The problem is that, due to how I've coded this, commands always have a valid config when they run (good), but that makes it hard to give clean gen-node output on first run. Someone runs nncp gen-node, it makes a default context, sets paths for stuff and then runs load_config, which makes the new config. Then gen-node can't yet tell if it existed before, so it will ask.
A problem for later; I'll likely stick a bool on context itself to indicate whether it was just created / existed before we ran load_config().
For now... Sleep...